Added support for pseudo classes, global stylesheet variables, more selector combinators. Added suffix/prefix to all category methods. Added more unit tests. Minor bug fixes and various refactorings.

+++ InterfaCSS/Model/ISSPointValue.m
@@ -83,7 +83,7 @@
 #pragma mark - NSObject overrides
 
 - (NSString*) description {
-    return [NSString stringWithFormat:@"ISSPointValue(%d - %@)", _type, NSStringFromCGPoint(_point)];
+    return [NSString stringWithFormat:@"ISSPointValue(%ld - %@)", (long)_type, NSStringFromCGPoint(_point)];
 }
 
 @end
